feat: Display and track activity status
Introduces `ActivityStatus` tracking and displays it in the UI. This change adds a new `activityStatusStats` to the dashboard for aggregated reporting and a new `ActivityStatusBadge` component for individual activity status visualization within the Practices Explorer.

The Activity form was also updated to correctly reflect the closed status of an activity.